+++ CHANGELOG.md
@@ -4,7 +4,12 @@ Both client libraries are pre-1.0, and they have separate versioning.
 
 ## Unreleased
 
+No unreleased changes.
+
+## modal-js/v0.3.24, modal-go/v0.0.24
+
 - Added `env` parameters to several methods, as a convenience for passing environment variables into Sandboxes, etc.
+- Added `Sandbox.getTags()` (JS) and `Sandbox.GetTags()` (Go).
 
 ## modal-js/v0.3.23, modal-go/v0.0.23